Fixes for Windows

+++ log/src/SqlStatement_TEST.cc
@@ -29,7 +29,7 @@ TEST(SqlParameter, Construct)
   EXPECT_EQ(transport::log::SqlParameter::ParamType::NULL_TYPE,
       transport::log::SqlParameter(nullptr).Type());
   EXPECT_EQ(transport::log::SqlParameter::ParamType::INTEGER,
-      transport::log::SqlParameter(42L).Type());
+      transport::log::SqlParameter(static_cast<int64_t>(42)).Type());
   EXPECT_EQ(transport::log::SqlParameter::ParamType::REAL,
       transport::log::SqlParameter(3.14159).Type());
   EXPECT_EQ(transport::log::SqlParameter::ParamType::TEXT,
@@ -45,7 +45,7 @@ TEST(SqlParameter, Set)
   EXPECT_EQ(transport::log::SqlParameter::ParamType::TEXT, param.Type());
   param.Set(3.14159);
   EXPECT_EQ(transport::log::SqlParameter::ParamType::REAL, param.Type());
-  param.Set(42L);
+  param.Set(static_cast<int64_t>(42));
   EXPECT_EQ(transport::log::SqlParameter::ParamType::INTEGER, param.Type());
   param.Set(nullptr);
   EXPECT_EQ(transport::log::SqlParameter::ParamType::NULL_TYPE, param.Type());
@@ -58,7 +58,7 @@ TEST(SqlParameter, QueryText)
   transport::log::SqlParameter param;
   param.Set("Hello World!");
   EXPECT_EQ(std::string("Hello World!"), *param.QueryText());
-  param.Set(42L);
+  param.Set(static_cast<int64_t>(42));
   EXPECT_EQ(nullptr, param.QueryText());
   param.Set(3.14159);
   EXPECT_EQ(nullptr, param.QueryText());
@@ -73,8 +73,8 @@ TEST(SqlParameter, QueryInteger)
   transport::log::SqlParameter param;
   param.Set("Hello World!");
   EXPECT_EQ(nullptr, param.QueryInteger());
-  param.Set(42L);
-  EXPECT_EQ(42L, *param.QueryInteger());
+  param.Set(static_cast<int64_t>(42));
+  EXPECT_EQ(static_cast<int64_t>(42), *param.QueryInteger());
   param.Set(3.14159);
   EXPECT_EQ(nullptr, param.QueryInteger());
   param.Set(nullptr);
@@ -88,7 +88,7 @@ TEST(SqlParameter, QueryReal)
   transport::log::SqlParameter param;
   param.Set("Hello World!");
   EXPECT_EQ(nullptr, param.QueryReal());
-  param.Set(42L);
+  param.Set(static_cast<int64_t>(42));
   EXPECT_EQ(nullptr, param.QueryReal());
   param.Set(3.14159);
   EXPECT_DOUBLE_EQ(3.14159, *param.QueryReal());
@@ -113,7 +113,7 @@ TEST(SqlParameter, CopyAssignment)
   transport::log::SqlParameter param_orig;
   param_orig.Set("Hello World!");
   transport::log::SqlParameter param;
-  param.Set(42L);
+  param.Set(static_cast<int64_t>(42));
   param = param_orig;
   ASSERT_NE(nullptr, param_orig.QueryText());
   ASSERT_NE(nullptr, param.QueryText());
@@ -128,7 +128,7 @@ TEST(SqlParameter, MoveAssignment)
   transport::log::SqlParameter param_orig;
   param_orig.Set("Hello World!");
   transport::log::SqlParameter param;
-  param.Set(42L);
+  param.Set(static_cast<int64_t>(42));
   param = std::move(param_orig);
   ASSERT_NE(nullptr, param.QueryText());
   EXPECT_EQ(std::string("Hello World!"), *param.QueryText());
